Modularized rust code
